EDX-5843 Move bootstrap/utils/http.go logic to rest package
